Graphics upgrade + bugs fixes
- post processing filter added
- burning color for bonuses
- fix on speed animation when getting a speed disease
- speed increment changed to 0.5

+++ Assets/Scripts/ScriptableObjects/Bonus/BonusSpeed.cs
@@ -3,7 +3,7 @@
 [CreateAssetMenu(menuName = "ScriptableAction/Bonus/BonusSpeed", fileName = "BonusSpeed")]
 public class BonusSpeed : BonusBehaviour<PlayerEntity>
 {
-    public int speedModifier;
+    public float speedModifier;
     public override void PerformAction(PlayerEntity playerEntity)
     {
        playerEntity.playerMovement.RpcModifySpeed(speedModifier);
